wsb-etf-sentinel/pyproject.toml
[project]
name = "wsb-etf-sentinel"
version = "1.0.0"
description = "Daily pipeline: scrape r/wallstreetbets, run sentiment analysis, compute ETF composition"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9,<3",
    "google-generativeai>=0.8,<1",
    "requests>=2.31,<3",
    "yfinance>=0.2,<1",
    "python-dotenv>=1.0,<2",
    "flask>=3.0,<4",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
